refactor: :art: modify directory
