+++ .storybook/main.js
@@ -1,15 +1,16 @@
 module.exports = {
-    stories: [
-        '../components/**/*.stories.@(ts|tsx)',
-        '../pages/**/*.stories.@(ts|tsx)',
-    ],
-    addons: [
-        '@storybook/addon-links',
-        '@storybook/addon-essentials',
-        '@storybook/addon-interactions',
-    ],
-    features: {
-        storyStoreV7: true,
-    },
-    framework: '@storybook/react',
+  stories: [
+    '../components/**/*.stories.@(ts|tsx)',
+    '../domains/**/*.stories.@(ts|tsx)',
+    '../pages/**/*.stories.@(ts|tsx)',
+  ],
+  addons: [
+    '@storybook/addon-links',
+    '@storybook/addon-essentials',
+    '@storybook/addon-interactions',
+  ],
+  features: {
+    storyStoreV7: true,
+  },
+  framework: '@storybook/react',
 };
