+++ pages/user/signin.tsx
@@ -1,0 +1,7 @@
+import Social from 'domains/user/signin/Social';
+
+function SignIn() {
+  return <Social />;
+}
+
+export default SignIn;
