+++ tsconfig.paths.json
@@ -14,6 +14,8 @@
         "@constants/*": ["constants/*"],
         "@constants": ["constants/index"],
         "@components/*": ["components/*"],
+        "@domains/*": ["domains/*"],
+        "@states/*": ["states/*"],
       }
     }
   }
